fix: Handle parent directory as destination (#69)
I'm open to the suggestion that I'm just doing this wrong :), but: 

The project I'm working with has multiple submodules. One of these modules (two subdirectories deep) we would like to publish with this tool to the gh-pages branch, but to the top-level directory on the gh-pages branch.

E.g.: 
```
ember github-pages:commit --message "Update gh-pages release" --destination ../../
```

This fails in two places: 
* rm -rf ../.. is problematic; 
* mkdir ../.. complains that it's creating a directory that already exists

To fix this I've added a regex that allows this case to be handled by the first condition of the branch.